[all] Add support for fault types in litmus tests postconditions
This change adds support for specifying the type of fault in litmus
tests postconditions. The allowed fault types depend on architecture
used.

Fault atoms in postconditions are allowed to not specify a fault type
in which case they are compared against against Faults of all
types. For example, an MMU:Translation fault in P0:L0 location x with
validate the atom Fault(P0:L0,x) in the same as the atom
Fault(P0:L0,x,MMU:Translation).

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>