[herd] Clarify the definition of same-loc in aarch64.cat
This change implements AARCH-23766 and AARCH-24727 as published in the
Arm Architecture Reference Manual for A-profile architecture: Known
issues in Issue L.b.

The relation "Same Location" which is meant to apply between two
Effects that are either:

- To the same physical address when the address
translation for these two Effects has a valid output address, or

- Share the same low-order bits when the translation for at least one
of these Effects does not have a valid output address.

This change is to facilitate the transliteration of Same Location and
it not meant to change the intent of the architecture.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>