[tools] Fix miaou to handle the intervening function
There are cat clauses that expect that there is no intervening Effect
in a specific relation. Add support for transliterating it with
miaou7.

For example:

let rel = intervening(M, po & loc)

will print latex that corresponds to the following text:

There is a rel between E1 and E2 if all of the following apply:

- E~1~ appears in program order before E~3~.

- E~1~ and E~3~ are to the same Physical Address.

- E~3~ is a Memory Effect.

- E~3~ appears in program order before E~1~.

- E~3~ and E~1~ are to the same Physical Address.

Signed-off-by: Nikos Nikoleris <nikos.nikoleris@arm.com>